+++ .gitignore
@@ -28,7 +28,6 @@ project/.sbtserver
 
 # Scala-IDE specific
 .scala_dependencies
-.worksheet
 *.sc
 *.iml
