+++ src/main/scala/de/htwg/se/muehle/Muehle.scala
@@ -1,6 +1,6 @@
 package de.htwg.se.muehle
 
-import de.htwg.se.muehle.controller.Controller
+import controller.controllerComponent.controllerBaseImpl.Controller
 import de.htwg.se.muehle.aview.Tui
 import model.gridComponent.gridBaseImpl.Grid
 import model.playerComponent.Player
