+++ src/main/scala/de/htwg/se/muehle/controller/controllerComponent/IController.scala
@@ -1,0 +1,9 @@
+package controller.controllerComponent
+
+trait IController {
+
+  def newGame():Unit
+  def gridToString: String
+  def placeStone(pos:Int):Unit
+  def moveStone(src:Int, pos:Int):Unit
+}
