+++ src/main/scala/de/htwg/se/muehle/controller/controllerComponent/controllerBaseImpl/ControllerStateStatusPlaced.scala
@@ -1,4 +1,4 @@
-package de.htwg.se.muehle.controller
+package controller.controllerComponent.controllerBaseImpl
 
 class ControllerStateStatusPlaced extends ControllerStateTemplate {
