+++ src/main/scala/de/htwg/se/muehle/controller/controllerComponent/controllerBaseImpl/ControllerStsteStatusMoved.scala
@@ -1,4 +1,4 @@
-package de.htwg.se.muehle.controller
+package controller.controllerComponent.controllerBaseImpl
 
 class ControllerStsteStatusMoved extends ControllerStateTemplate {
